+++ src/components/common/CreateButton/index.tsx
@@ -1,0 +1,13 @@
+import React from "react";
+import { PencilIcon } from "../../Icons/BoardIcons";
+import { Button } from "./style";
+
+const CreateButton = () => {
+  return (
+    <Button>
+      <PencilIcon />글 작성하기
+    </Button>
+  );
+};
+
+export default CreateButton;
