+++ src/components/common/CreateButton/style.ts
@@ -1,7 +1,7 @@
 import styled from "styled-components";
-import COLOR from "./../../../constants/color";
+import COLOR from "../../../constants/color";
 
-const CreateButton = styled.button`
+const Button = styled.button`
   color: ${COLOR.GREEN4};
   background: ${COLOR.WHITE};
   padding: 0.5rem 1rem;
@@ -10,4 +10,4 @@ const CreateButton = styled.button`
   align-items: center;
 `;
 
-export { CreateButton };
+export { Button };
