+++ src/components/common/PageTitle/index.tsx
@@ -1,0 +1,8 @@
+import React from "react";
+import { Title } from "./style";
+
+const PageTitle = ({ children }: { children: React.ReactNode }) => (
+  <Title>{children}</Title>
+);
+
+export default PageTitle;
